requirements.txt
django
djangorestframework
frozendict
pytest-django
psycopg2-binary
pytz
pytest-freezegun